backend/requirements.txt
anyio==3.5.0
attrs==21.4.0
black==21.12b0
click==8.0.3
fastapi==0.73.0
Flask==2.0.2
Flask-Cors==3.0.10
grpcio==1.43.0
grpcio-tools==1.43.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
protobuf==3.19.3
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
tomli==1.2.3
typing-extensions==4.0.1
Werkzeug==2.0.2
